// CSE-110-Lab-3/my-app-lab3/src/types.ts
export enum Label {
    personal = "personal",
    study = "study",
    work = "work",
    other = "other",
 }
 
 export type Note = {
    id: number;
    title: string;
    content: string;
    label: Label;
    favorite: boolean;
 };

export type GroceryItem = { name: string; isPurchased: boolean };